bridge/requirements.txt
confluent-kafka
celery[redis]
python-dotenv